fix: include pip mirror helper in server build
